backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
pydantic==2.5.3
pydantic-settings==2.1.0
alembic==1.13.1
python-dotenv==1.0.0
python-multipart==0.0.6
celery[redis]==5.3.6
redis==5.0.1
pandas==2.2.0
httpx==0.27.0
python-dateutil==2.8.2
